Inventory part 2
Improved method of writing json
Finally fixed getting correct inventory values
