+++ app.py
@@ -1,5 +1,6 @@
 ##### Importing method from methods.py (all methods must be stored there or other files, not here)
 from methods import welcome
+from character import create_json
 
-
+create_json()
 welcome()
